src/components/atoms/image.tsx
import React, { useMemo } from 'react';
import { ImageSourcePropType, Image as RNImage } from 'react-native';

import { Image as ExpoImage, ImageProps as ExpoImageProps } from 'expo-image';

interface CustomImageProps extends Omit<ExpoImageProps, 'source'> {
  source: ImageSourcePropType;
  calculateAspectRatio?: boolean;
}

const BLURHASH =
  '|rF?hV%2WCj[ayj[a|j[az_NaeWBj@ayfRayfQfQM{M|azj[azf6fQfQfQIpWXofj[ayj[j[fQayWCoeoeaya}j[ayfQa{oLj?j[WVj[ayayj[fQoff7azayj[ayj[j[ayofayayayj[fQj[ayayj[ayfjj[j[ayjuayj[';

/**
 * A wrapper to use the Expo Image component and calculate aspect ratio if needed
 * @param calculateAspectRatio - If true, the aspect ratio of the image will be calculated
 * @param style - The style of the image
 * @param placeholder - The placeholder of the image
 * @param source - The source of the image
 * @returns
 */
export default function Image({
  calculateAspectRatio = false,
  style,
  placeholder,
  source,
  ...props
}: CustomImageProps) {
  const aspectRatioStyle = useMemo(() => {
    if (calculateAspectRatio && source) {
      const { width, height } = RNImage.resolveAssetSource(source);
      return { aspectRatio: width / height };
    }
    return {};
  }, [calculateAspectRatio]);

  return (
    <ExpoImage
      {...props}
      placeholder={placeholder ?? BLURHASH}
      style={[style, aspectRatioStyle]}
      source={source}
    />
  );
}
